tests/test_api_endpoints.py
"""
Tests para API endpoints

Cobre:
- REST API responses
- Status codes
- Autenticação de endpoints
- Validação de dados
"""

import pytest
from rest_framework import status

pytestmark = pytest.mark.django_db


class TestAPIAuthentication:
    """Tests para autenticação em endpoints"""

    def test_unauthenticated_access_denied(self, api_client):
        """Testa que acesso não autenticado é negado"""
        # Placeholder - requer endpoints
        pass

    def test_authenticated_access_allowed(self, authenticated_client, user):
        """Testa que acesso autenticado é permitido"""
        # Placeholder - requer endpoints
        pass

    def test_invalid_token_denied(self, api_client):
        """Testa que token inválido é negado"""
        # Placeholder - requer endpoints
        pass


class TestAPIValidation:
    """Tests para validação de dados em API"""

    def test_invalid_data_returns_400(self, authenticated_client):
        """Testa que dados inválidos retornam 400"""
        # Placeholder - requer endpoints
        pass

    def test_missing_required_fields_returns_400(self, authenticated_client):
        """Testa que campos obrigatórios faltando retornam 400"""
        # Placeholder - requer endpoints
        pass

    def test_valid_data_returns_201(self, authenticated_client):
        """Testa que dados válidos retornam 201"""
        # Placeholder - requer endpoints
        pass


class TestAPIPagination:
    """Tests para paginação de API"""

    def test_paginated_list_response(self, authenticated_client):
        """Testa resposta paginada"""
        # Placeholder - requer endpoints
        pass

    def test_pagination_params(self, authenticated_client):
        """Testa parâmetros de paginação"""
        # Placeholder - requer endpoints
        pass


class TestAPIFiltering:
    """Tests para filtragem de API"""

    def test_filter_by_status(self, authenticated_client):
        """Testa filtragem por status"""
        # Placeholder - requer endpoints
        pass

    def test_filter_by_date_range(self, authenticated_client):
        """Testa filtragem por range de datas"""
        # Placeholder - requer endpoints
        pass

    def test_search_functionality(self, authenticated_client):
        """Testa funcionalidade de busca"""
        # Placeholder - requer endpoints
        pass
