requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-tenants==3.5.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
channels==4.0.0
daphne==4.0.0

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Authentication & Authorization
djangorestframework-simplejwt==5.3.2
django-oauth-toolkit==2.3.0

# Payment Gateways
stripe==7.8.0
paypalrestsdk==1.7.4
razorpay==1.4.1

# Storage
django-storages==1.14.2
boto3==1.34.0

# API & Serialization
requests==2.31.0
python-dateutil==2.8.2

# Data Validation & Modeling
marshmallow==3.20.1
pydantic==2.5.2

# Monitoring & Logging
sentry-sdk==1.38.0

# Code Quality & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development Tools
ipython==8.18.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Utilities
pytz==2023.3.post1
Pillow==10.1.0
openpyxl==3.11.0
reportlab==4.0.7

# Timezone Support
django-timezone-field==6.1.0

# Notifications & Email
django-anymail==10.1
sendgrid==6.10.0

# File Processing
python-magic==0.4.27
pdf2image==1.16.3

# QR Code & Barcode
qrcode==7.4.2
python-barcode==0.14.0

# Biometric Integration (if needed)
pyfingerprint==0.0.1

# PWA & Progressive Web Apps
pwa==1.1.0
django-pwa==0.0.13
whitenoise==6.6.0

# Push Notifications
django-push-notifications==3.0.2
pywebpush==1.12.0

# Image Processing (for PWA icons)
pillow-heif==0.7.1

# Rate Limiting & Security
django-ratelimit==4.1.0
django-cors-headers==4.3.1

# CI/CD & Testing
pytest-xdist==3.5.0
pytest-timeout==2.2.0
